requirements.txt
langchain
openai
pytube
python-dotenv
requests